% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlotDS.R
\name{clusterPlotDS}
\alias{clusterPlotDS}
\title{Conducts k-means clustering of a given data set}
\usage{
clusterPlotDS(
  df.name,
  k,
  h,
  k_colors,
  color_labels_by_k,
  rect,
  main,
  xlab,
  ylab
)
}
\arguments{
\item{df.name}{is a string character of the data set}

\item{k}{specifies the number of clusters in which the tree should be cut}

\item{h}{specifies the height of a tree at which the tree should be cut}

\item{k_colors}{is a vector containing colors to be used for groups}

\item{color_labels_by_k}{is a logical value which colors the branches by group when k is not NULL}

\item{rect}{is a logical value which specifies whether to add a rectangle around groups when k is not NULL}

\item{main}{is the main title of the plot}

\item{xlab}{is a title for the x axis}

\item{ylab}{is a title for the y axis}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.kmeans} or default name \code{kmeans.newobj}
}
\description{
This function is similar to the native R function 'kmeans' from stats
}
\details{
The function calls the server-side function \code{kmeansDS} that computes the
k-means clustering of a data set (type 'data.frame' or 'matrix'). 
The function creates a new object on the server-side, which is of class 'kmeans'.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{kmeans.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
