% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutreeDS.R
\name{cutreeDS}
\alias{cutreeDS}
\title{Cuts the tree of a hclust object}
\usage{
cutreeDS(tree, k = NULL, h = NULL)
}
\arguments{
\item{tree}{is a string character specifying the name of the hclust object}

\item{k}{specifies the number of clusters in which the tree should be cut}

\item{h}{specifies the height of a tree at which the tree should be cut}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.cutree} or default name \code{cutree.newobj}
}
\description{
This function is similar to the native R function from stats
}
\details{
The function computes the clusters for a given number of clusters k or height h of the tree, and assigns the new object to the server-side.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{cutree.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
