% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustDS.R
\name{hclustDS}
\alias{hclustDS}
\title{Conducts hierarchical clustering of a given dissimilarity structure}
\usage{
hclustDS(diss, method)
}
\arguments{
\item{diss}{is a string character of the dissimilarity structure}

\item{method}{specifies the method for the calculation of the hierarchical clustering and can be either 'ward.D', 'ward.D2', 'single', 
'complete', 'average', 'mcquitty', 'median' or 'centroid'}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.hclust} or default name \code{hclust.newobj}
}
\description{
This function is similar to the native R function 'hclust' from stats
}
\details{
The function computes the hierarchical clustering of a dissimilarity structure (type 'dist').
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
