% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclustDS.R
\name{nbclustDS}
\alias{nbclustDS}
\title{Determines optimal number of clusters for dataset}
\usage{
nbclustDS(
  df.name,
  FUNcluster,
  method,
  diss,
  k.max,
  nboot,
  verbose,
  barfill,
  barcolor,
  linecolor,
  print.summary
)
}
\arguments{
\item{df.name}{is a string character of the data set and can be either a matrix or data frame}

\item{FUNcluster}{is a partitioning function}

\item{method}{the method to be used for estimating the optimal number of clusters. Possible values are "silhouette", "wss" and "gap_stat".}

\item{diss}{'dist' object as produced by ds.dist. If diss = NULL, dist(df.name) is computed with the default "euclidean" setting.}

\item{k.max}{maximum number of clusters to be considered (has to be at least 2)}

\item{nboot}{number of Monte Carlo bootstrap samples. This argument is only used for determining the number of clusters for gap statistics}

\item{verbose}{logical values, If TRUE, the result of progress is printed.}

\item{barfill}{fill color for bars}

\item{barcolor}{outline color for bars}

\item{linecolor}{color for lines}

\item{print.summary}{logical value. If TRUE, the optimal number of clusters are printed}
}
\value{
a ggplot2 image suggesting optimal number of clusters
}
\description{
This function is similar to the R function 'fviz_nbclust' from the factoextra package
}
\details{
The function uses partitioning methods to find optimal numbers of clusters for a given dataset.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
