% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelLcmDS2.R
\name{varSelLcmDS2}
\alias{varSelLcmDS2}
\title{Needs editing}
\usage{
varSelLcmDS2(
  df,
  num.clust,
  vbleSelec,
  crit.varsel,
  initModel,
  nbcores,
  nbSmall,
  iterSmall,
  nbKeep,
  iterKeep,
  tolKeep,
  num.iterations,
  initialRun_char_vect,
  colnames_char_vect,
  entries_per_study
)
}
\arguments{
\item{df}{is a string character of the data set}

\item{num.clust}{specifies the number of clusters for the computation}

\item{vbleSelec}{specifies the max. number of iterations allowed}

\item{crit.varsel}{relates to the number of random sets if clusters is a number and not a set of initial cluster centers}

\item{initModel}{refers to the algorithm of calculating the kmeans and can be either 'Hartigan-Wong', 'Lloyd', 'Forgy' or 'MacQueen'}

\item{nbcores}{is the name of the new object which is created with this function}

\item{nbSmall}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{iterSmall}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{nbKeep}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{iterKeep}{is a logical or integer specifying whether tracing information on the progress of the algorithm is procuded for the Hartigan-Wong algorithm}

\item{tolKeep}{represents the number at which point two successive models are defined to be converged; default is 1e-7}

\item{num.iterations}{the number of iterations for finding SLMA clusters in each respective datasource}

\item{initialRun_char_vect}{needs editing}

\item{colnames_char_vect}{needs editing}

\item{entries_per_study}{needs editing}
}
\description{
Needs editing
}
\details{
Needs editing
}
