% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelLcmDS3.R
\name{varSelLcmDS3}
\alias{varSelLcmDS3}
\title{Needs editing}
\usage{
varSelLcmDS3(df)
}
\arguments{
\item{df}{is a string character of the data set}
}
\description{
Needs editing
}
\details{
Needs editing
}
