% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelLcmDS4.R
\name{varSelLcmDS4}
\alias{varSelLcmDS4}
\title{Needs editing}
\usage{
varSelLcmDS4(
  num.clust,
  final_match_char,
  final_colnames_char = c("results_values", "Matching")
)
}
\arguments{
\item{num.clust}{specifies the number of clusters for the computation}

\item{final_match_char}{needs editing}

\item{final_colnames_char}{needs editing}
}
\description{
Needs editing
}
\details{
Needs editing
}
