% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlotDS.R
\name{clusterPlotDS}
\alias{clusterPlotDS}
\title{Plots hierarchical cluster}
\usage{
clusterPlotDS(
  df.name,
  k,
  h,
  k_colors,
  palette,
  show_labels = FALSE,
  color_labels_by_k
)
}
\arguments{
\item{df.name}{is a string character of the data set}

\item{k}{specifies the number of clusters in which the tree should be cut}

\item{h}{specifies the height of a tree at which the tree should be cut}

\item{k_colors}{is a vector containing colors to be used for groups}

\item{color_labels_by_k}{is a logical value which colors the branches by group when k is not NULL}

\item{rect}{is a logical value which specifies whether to add a rectangle around groups when k is not NULL}

\item{main}{is the main title of the plot}

\item{xlab}{is a title for the x axis}

\item{ylab}{is a title for the y axis}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.kmeans} or default name \code{kmeans.newobj}
}
\description{
This function is similar to the R function 'fviz_dend' from the factoextra package
}
\details{
The function computes the dendrogram without any labels to prevent any disclosures.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{kmeans.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
